requirements.txt
fastapi
uvicorn
python-multipart
pydantic
ollama

google-genai

openai-whisper
torch
torchaudio
pyttsx3

sentence-transformers
faiss-cpu
pypdf

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

python-dotenv
requests
playwright
ollama
google-genai
python-dotenv
requests
